move the flashcard types into a singular file, and create a union type for both the base flashcard and the firestore flashcard
