--- src/types/flashcard.ts
@@ -1,11 +1,0 @@
-// Base Flashcard type
-export interface Flashcard {
-  front: string;
-  back: string;
-}
-
-// Firestore Flashcard type extending the base Flashcard
-// with document id from Firebase
-export interface FirestoreFlashcard extends Flashcard {
-  id: string;
-}
